refractoring: move identifiers from Domain/Identifers to Fluent/Freeze
