+++ ArchUnitNET/Fluent/Slices/SliceRuleInitializer.cs
@@ -7,7 +7,6 @@
 
 using System;
 using ArchUnitNET.Domain;
-using ArchUnitNET.Domain.Identifiers;
 
 namespace ArchUnitNET.Fluent.Slices
 {
