+++ ArchUnitNET/Fluent/Slices/SlicesShould.cs
@@ -8,7 +8,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using ArchUnitNET.Domain;
-using ArchUnitNET.Domain.Identifiers;
+using ArchUnitNET.Fluent.Freeze;
 using StronglyConnectedComponents;
 
 namespace ArchUnitNET.Fluent.Slices
